evals/pyproject.toml
[project]
name = "aap-evals"
version = "0.1.0"
description = "AAP apply-engine evaluation suite — generates artifacts via Ollama, validates via Rust FFI"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "typer>=0.24",
    "rich>=14",
    "pydantic-ai-slim[openai,google]>=1.0",
    "pydantic>=2.0",
    "ragas>=0.4",
    "aiohttp>=3.11",
    "maturin>=1.8",
    "cffi",
    "tenacity>=9.0",
]

[project.scripts]
aap-evals = "aap_evals.cli:app"

[build-system]
requires = ["maturin>=1.8"]
build-backend = "maturin"

[tool.maturin]
manifest-path = "../Cargo.toml"
features = ["python"]
module-name = "aap_evals.aap"
python-source = "src"
